[SYCL] fix types

+++ sycl/include/sycl/detail/spirv.hpp
@@ -501,7 +501,7 @@ AtomicMax(multi_ptr<T, AddressSpace, IsDecorated> MPtr, memory_scope Scope,
 // - The Intel SPIR-V extension natively supports all arithmetic types.
 //   However, OpenCL extension natively supports float vectors,
 //   integer vectors, half scalar and double scalar.
-//   For double,long long and half vectors we perform emulation with scalar
+//   For double, long and half vectors we perform emulation with scalar
 //   version.
 // - The CUDA shfl intrinsics do not support vectors, and we use the _i32
 //   variants for all scalar types
@@ -510,7 +510,7 @@ AtomicMax(multi_ptr<T, AddressSpace, IsDecorated> MPtr, memory_scope Scope,
 template <typename T>
 struct TypeIsProhibitedForShuffleEmulation
     : bool_constant<std::is_same_v<vector_element_t<T>, double> ||
-                    std::is_same_v<vector_element_t<T>, long long> ||
+                    std::is_same_v<vector_element_t<T>, long> ||
                     std::is_same_v<vector_element_t<T>, half>> {};
 
 template <typename T>
